+++ core/modules/dblog/src/Plugin/rest/resource/DBLogResource.php
@@ -3,7 +3,7 @@
 namespace Drupal\dblog\Plugin\rest\resource;
 
 use Drupal\Core\Plugin\ResourceBase;
-use Drupal\rest\ResourceResponse;
+use Drupal\Core\Response\ResourceResponse;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
@@ -28,7 +28,7 @@ class DBLogResource extends ResourceBase {
    * @param int $id
    *   The ID of the watchdog log entry.
    *
-   * @return \Drupal\rest\ResourceResponse
+   * @return \Drupal\Core\Response\ResourceResponse
    *   The response containing the log entry.
    *
    * @throws \Symfony\Component\HttpKernel\Exception\NotFoundHttpException
